Move ColourBundle to internal class. Remove unused icon file.
