+++ ALsSoundSwitcher_Frontend/ALsSoundSwitcher/ALsSoundSwitcher.csproj
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="..\packages\Costura.Fody.5.7.0\build\Costura.Fody.props" Condition="Exists('..\packages\Costura.Fody.5.7.0\build\Costura.Fody.props')" />
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
@@ -77,7 +77,6 @@
     </Compile>
     <Compile Include="Source\Logic\DeviceUtils.cs" />
     <Compile Include="Source\Logic\DeviceUpdateCallbacks.cs" />
-    <Compile Include="Source\Theming\ColourBundle.cs" />
     <Compile Include="Source\Theming\ColourPack.cs" />
     <Compile Include="Source\Theming\MenuStripColorTable.cs" />
     <Compile Include="Source\Theming\ThemeCreator.cs">
