--- ALsSoundSwitcher_Frontend/ALsSoundSwitcher/Source/Theming/ColourBundle.cs
@@ -1,18 +1,0 @@
-﻿using System.Drawing;
-
-namespace ALsSoundSwitcher
-{
-  public class ColourBundle
-  {
-    public ColourBundle(Color colour, string jsonKey, Bitmap mask)
-    {
-      Colour = colour;
-      JsonKey = jsonKey;
-      Mask = mask;
-    }
-
-    public Color Colour;
-    public string JsonKey;
-    public Bitmap Mask;
-  }
-}
